Refactor API return type by removing AccountRequestCreateResponseData
